tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/fallback.multiple.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: fallback.multiple</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>fallback.multiple</h1>
<p class="desc">Fallback content with multiple elements</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL</p><p class="fallback">FAIL</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Fallback content with multiple elements");
_addTest(function(canvas, ctx) {

_assertSame(canvas.childNodes.length, 2, "canvas.childNodes.length", "2");


});
</script>
